Unified drawing methods, GIMP plugin resizing
The drawing methods for individual tiles and the map as a whole now call
a unified, per-(x,y) coordinate method as a backend. Now when modifying
how things are drawn, only this single method needs updating. I also
modified the GIMP plugin to remove the built-in update API since it was
leading to an annoying error message. The plugin also reads map
dimension info from the input map and resizes the image automatically,
which is a nice time-saving feature.